Run dropbox commands as dropbox owner (#3)
Dropbox status and start commands were always showing up as 'changed',
even Dropbox was known to be running.  Debugging showed that it was
because the commands were being run as root, but Dropbox was (correctly)
running as dropbox_owner.

Run both commands as dropbox_owner to avoid spurious 'changed' statuses
while deploying to a running instance.

Note that on a new deployment, the start comand will repeatedly show up
as 'changed' because to start a new instance you have to run 'dropbox
start -i' and then authenticate to Dropbox.  Automating that with
ansible would likely be difficult if even possible.